webapp/ImageProcessor/requirements.txt
numpy==1.23.5
opencv-python-headless==4.6.0.66
tensorflow-cpu==2.9.3
